+++ packages/lib/src/typeChecking/tProp.ts
@@ -67,7 +67,6 @@ export function tProp(defaultValue: boolean, options?: ModelPropOptions): Option
  * ```
  *
  * @typeparam TType Type checker type.
- *
  * @param type Type checker.
  * @param defaultFn Default value generator function.
  * @param options Model property options.
@@ -91,7 +90,6 @@ export function tProp<TType extends AnyType>(
  * ```
  *
  * @typeparam TType Type checker type.
- *
  * @param type Type checker.
  * @param defaultValue Default value generator function.
  * @param options Model property options.
@@ -113,7 +111,6 @@ export function tProp<TType extends AnyType>(
  * ```
  *
  * @typeparam TType Type checker type.
- *
  * @param type Type checker.
  * @param options Model property options.
  * @returns