+++ packages/lib/src/utils/index.ts
@@ -1,3 +1,4 @@
+import * as mobx from "mobx"
 import {
   IObservableArray,
   isObservableArray,
@@ -8,8 +9,6 @@ import {
   ObservableSet,
 } from "mobx"
 import { PrimitiveValue } from "./types"
-
-const { makeObservable } = require("mobx")
 
 /**
  * A mobx-keystone error.
@@ -403,12 +402,22 @@ export function lazy<V>(valueGen: () => V): () => V {
   }
 }
 
+// just to ensure import * is kept properly
+/**
+ * @ignore
+ * @internal
+ */
+export const mobx6 = {
+  // eslint-disable-next-line no-useless-concat
+  makeObservable: (mobx as any)["makeObservable" + ""] as typeof mobx["makeObservable"],
+}
+
 /**
  * @ignore
  * @internal
  */
 export function getMobxVersion(): number {
-  if (makeObservable) {
+  if (mobx6.makeObservable!) {
     return 6
   } else {
     return 5